# Add better-performance read-entire-record methods