… CMakeLists.txt
… src/stim/io/measure_record_reader.h
… src/stim/io/measure_record_reader.test.cc
… src/stim/io/sparse_shot.test.cc
… src/stim/main_namespaced.test.cc
… src/stim/mem/simd_bits.cc
… src/stim/mem/simd_bits.h
… src/stim/mem/simd_bits_range_ref.h
… src/stim/mem/simd_bits_range_ref.test.cc