- Add a dense (into bit packed memory) and sparse (into vector of hits) variants
- Add SparseShot struct for reading sparse data
- Fix various readers not knowing about the expected layout of data they were reading
- Add benchmarks for read performance
- Add deprecation warning to `--prepend_observables`
- Add prefix_ref method to simd_bits
- Move simd-bit-sizing methods into stim:: namespace

```
[....................*....................]  80 us (vs  80 us) (120 MBits/s) read_01_dense_per10
[....................*....................]  51 us (vs  57 us) ( 19 MPops/s) read_01_sparse_per10
[....................*....................] 5.0 us (vs 5.6 us) (1.9 GBits/s) read_b8_dense_per10
[....................*....................] 7.8 us (vs 8.0 us) (120 MPops/s) read_b8_sparse_per10
[....................*....................]  22 us (vs  22 us) (430 MBits/s) read_hits_dense_per10
[....................*....................] 2.7 us (vs 2.8 us) (3.5 GBits/s) read_hits_dense_per100
[....................*....................]  22 us (vs  25 us) ( 44 MPops/s) read_hits_sparse_per10
[....................|*...................] 2.6 us (vs 3.4 us) ( 37 MPops/s) read_hits_sparse_per100
[....................*....................]  30 us (vs  30 us) (320 MBits/s) read_dets_dense_per10
[....................*....................] 3.6 us (vs 3.6 us) (2.7 GBits/s) read_dets_dense_per100
[....................|*...................]  27 us (vs  33 us) ( 36 MPops/s) read_dets_sparse_per10
[....................*....................] 3.2 us (vs 3.6 us) ( 30 MPops/s) read_dets_sparse_per100
[...................*|....................] 7.1 us (vs 6.3 us) (1.4 GBits/s) read_r8_dense_per10
[....................*....................] 1.2 us (vs 1.3 us) (7.9 GBits/s) read_r8_dense_per100
[....................*....................] 4.6 us (vs 4.4 us) (210 MPops/s) read_r8_sparse_per10
[....................*....................] 1.0 us (vs 1.0 us) ( 93 MPops/s) read_r8_sparse_per100
```